Split out SlackIntegrationCommander
