+++ kifi-backend/modules/shoebox/app/com/keepit/controllers/website/SlackController.scala
@@ -13,7 +13,7 @@ import com.keepit.model.LibrarySpace.{ OrganizationSpace, UserSpace }
 import com.keepit.model._
 import com.keepit.shoebox.controllers.OrganizationAccessActions
 import com.keepit.slack.models._
-import com.keepit.slack.{ SlackTeamCommander, SlackAuthenticatedAction, SlackClient, SlackCommander }
+import com.keepit.slack._
 import play.api.libs.json._
 import play.api.mvc.Result
 
@@ -24,6 +24,7 @@ import scala.util.{ Failure, Success }
 class SlackController @Inject() (
     slackClient: SlackClient,
     slackCommander: SlackCommander,
+    slackIntegrationCommander: SlackIntegrationCommander,
     slackTeamCommander: SlackTeamCommander,
     libraryAccessCommander: LibraryAccessCommander,
     deepLinkRouter: DeepLinkRouter,
@@ -75,19 +76,19 @@ class SlackController @Inject() (
     action match {
       case SetupLibraryIntegrations => (Library.decodePublicId(data), slackAuth.incomingWebhook) match {
         case (Success(libId), Some(webhook)) =>
-          slackCommander.setupIntegrations(userId, libId, webhook, slackIdentity)
+          slackIntegrationCommander.setupIntegrations(userId, libId, webhook, slackIdentity)
           Future.successful(redirectToLibrary(libId, showSlackDialog = true))
         case _ => Future.successful(BadRequest("invalid_library_id"))
       }
       case TurnOnLibraryPush => (LibraryToSlackChannel.decodePublicId(data), slackAuth.incomingWebhook) match {
         case (Success(integrationId), Some(webhook)) =>
-          val libraryId = slackCommander.turnOnLibraryPush(integrationId, webhook, slackIdentity)
+          val libraryId = slackIntegrationCommander.turnOnLibraryPush(integrationId, webhook, slackIdentity)
           Future.successful(redirectToLibrary(libraryId, showSlackDialog = true))
         case _ => Future.successful(BadRequest("invalid_integration_id"))
       }
       case TurnOnChannelIngestion => SlackChannelToLibrary.decodePublicId(data) match {
         case Success(integrationId) =>
-          val libraryId = slackCommander.turnOnChannelIngestion(integrationId, slackIdentity)
+          val libraryId = slackIntegrationCommander.turnOnChannelIngestion(integrationId, slackIdentity)
           Future.successful(redirectToLibrary(libraryId, showSlackDialog = true))
         case _ => Future.successful(BadRequest("invalid_integration_id"))
       }
@@ -151,7 +152,7 @@ class SlackController @Inject() (
           }.toSet
         }
         if (libraryAccessCommander.ensureUserCanWriteTo(request.userId, libsUserNeedsToWriteTo)) {
-          slackCommander.modifyIntegrations(SlackIntegrationModifyRequest(request.userId, libToSlackMods, slackToLibMods)).map { response =>
+          slackIntegrationCommander.modifyIntegrations(SlackIntegrationModifyRequest(request.userId, libToSlackMods, slackToLibMods)).map { response =>
             Ok(Json.toJson(response))
           }.recover {
             case fail: LibraryFail => fail.asErrorResponse
@@ -168,7 +169,7 @@ class SlackController @Inject() (
       case JsSuccess(dels, _) =>
         val libToSlackDels = dels.collect { case Left(ltsId) => LibraryToSlackChannel.decodePublicId(ltsId).get }.toSet
         val slackToLibDels = dels.collect { case Right(stlId) => SlackChannelToLibrary.decodePublicId(stlId).get }.toSet
-        slackCommander.deleteIntegrations(SlackIntegrationDeleteRequest(request.userId, libToSlackDels, slackToLibDels)).map { response =>
+        slackIntegrationCommander.deleteIntegrations(SlackIntegrationDeleteRequest(request.userId, libToSlackDels, slackToLibDels)).map { response =>
           Ok(Json.toJson(response))
         }.recover {
           case fail: LibraryFail => fail.asErrorResponse
